refactor(ws-client): extract shared Thread/WiFi topology derivation
Move the pure network-topology derivation (neighbor/route table parsing,
edge-pair building, unknown/border-router classification, diagnostic
route64/childTable merging, WiFi diagnostics) out of the dashboard into
@matter-server/ws-client so the server can reuse it.

- Functions take a structural TopologySourceNode (Pick of MatterNodeData)
  instead of the dashboard MatterNode class, so the pipeline runs in both the
  browser and Node.
- Drop the DOM-derived `signalColor` field from ThreadConnection; callers
  derive the color from `signalLevel` at render time, leaving the pipeline
  free of theme/DOM dependencies.
- network-types.ts / network-utils.ts keep the render, color and side-panel
  (NodeConnection) helpers and re-export the moved symbols, so existing
  dashboard imports are unchanged.
- Add TopologyUtilsTest.ts: first direct coverage of the derivation pipeline.

No behaviour change to the dashboard.
